backend/requirements.txt
Django==5.0.2
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.1.0
drf-yasg==1.21.7